+++ src/util/constants.ts
@@ -1,12 +1,15 @@
 export const API_URL = {
-  prod: 'https://api.vocdoni.net/v2',
   dev: 'https://api-dev.vocdoni.net/v2',
+  stg: 'https://api-stg.vocdoni.net/v2',
+  prod: 'https://api.vocdoni.net/v2',
 };
 
 export const FAUCET_URL = {
   dev: 'https://faucet-azeno.vocdoni.net/faucet/vocdoni/dev',
+  stg: 'https://faucet-azeno.vocdoni.net/faucet/vocdoni/stg',
 };
 
 export const FAUCET_AUTH_TOKEN = {
   dev: '158a58ba-bd3e-479e-b230-2814a34fae8f',
+  stg: '158a58ba-bd3e-479e-b230-2814a34fae8f',
 };